requirements.txt
pytumblr2
python-dotenv
